+++ osrs-cache-service/app/db/models/npc.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from sqlalchemy import Boolean, Integer, String, UniqueConstraint
+from sqlalchemy import JSON, Boolean, Integer, String, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import Base
@@ -19,3 +19,13 @@ class Npc(Base):
     category: Mapped[int | None] = mapped_column(Integer, nullable=True)
     interactable: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True)
     minimap_visible: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True)
+    model_ids: Mapped[list[int] | None] = mapped_column(JSON, nullable=True)
+    chathead_model_ids: Mapped[list[int] | None] = mapped_column(JSON, nullable=True)
+    actions: Mapped[list[str | None] | None] = mapped_column(JSON, nullable=True)
+    color_find: Mapped[list[int] | None] = mapped_column(JSON, nullable=True)
+    color_replace: Mapped[list[int] | None] = mapped_column(JSON, nullable=True)
+    texture_find: Mapped[list[int] | None] = mapped_column(JSON, nullable=True)
+    texture_replace: Mapped[list[int] | None] = mapped_column(JSON, nullable=True)
+    varbit_id: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    varp_index: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    configs: Mapped[list[int | None] | None] = mapped_column(JSON, nullable=True)